V.1.4
__V1.4 Changes__ @everyone
• Added a history checker to view profitable and fast selling items (requires new module - ext.buttons)
• Updated to using just one main file however you do need to update the `Login_details/preferences.py` and add a line with either `cli_login = True/False`
• Improved the logging output to make it not break
• Subclassed Bot in the main file and running the steam side of the bot in executor - this allows you to run the bot using pm2
• Used a bit of regex magic to make the profit graphing simpler and much better performance wise
• Fixed bugs with user messages not showing up

• General code clean up, some embed changes and the usual boring programming house-keeping

+++ Cogs/loader.py
@@ -1,7 +1,6 @@
 from asyncio import sleep
 from datetime import datetime
-from os import remove, execv
-from sys import argv, executable
+from os import remove
 
 from discord import Activity, ActivityType
 from discord.ext import commands
@@ -10,7 +9,7 @@ from Login_details import preferences, sensitive_details
 
 
 def __version__():
-    return '1.3.3'
+    return '1.4'
 
 
 class Loader(commands.Cog):
@@ -20,13 +19,6 @@ class Loader(commands.Cog):
     def __init__(self, bot):
         """Setting all of your bot vars to be used by other cogs/commands"""
         self.bot = bot
-        bot.username = sensitive_details.username
-        bot.password = sensitive_details.password
-        try:
-            bot.secrets = sensitive_details.secrets
-        except AttributeError:
-            pass
-
         bot.bot64id = preferences.bots_steam_id
         bot.color = int(preferences.embed_colour, 16)
         bot.templocation = preferences.path_to_temp
@@ -49,32 +41,33 @@ class Loader(commands.Cog):
         """Setting your status, printing your bot's id to send to me,
         seeing if the bot was restarted and stored your number of trades,
         finally checking if you are logged onto steam"""
-        info = await self.bot.application_info()
-        self.bot.owner = info.owner
-        await self.bot.change_presence(activity=Activity(name=f'{self.bot.owner.name}\'s trades | V{__version__()}',
-                                                    type=ActivityType.watching))
-        print(f'{"-" * 30}\n{self.bot.user.name} is ready')
-        print(f'Send this id: "{self.bot.user.id}" to Gobot1234 to add your bot to the server to use the custom emojis')
-        print(f'This is: Version {__version__()}')
-        while self.bot.logged_on is False:
-            if self.bot.cli_login:
-                await self.bot.owner.send('You aren\'t currently logged into your Steam account\nTo do that type in '
-                                     'your 2FA code into the console.')
-            await sleep(60)
-        print('-' * 30)
-        try:
-            self.bot.trades = int(open('trades.txt', 'r').read())
-            channel = self.bot.get_channel(int(open('channel.txt', 'r').read()))
-            if channel is not None:
-                async for m in channel.history(limit=2):
-                    if m.author == self.bot.user:
-                        await m.delete()
-                await channel.send('Finished restarting...', delete_after=10)
-            remove('channel.txt')
-            remove('trades.txt')
-        except FileNotFoundError:
-            pass
-        await self.bot.owner.send('I\'m online both Steam and Discord dealing with your Steam messages')
+        if self.bot.first:
+            info = await self.bot.application_info()
+            self.bot.owner = info.owner
+            await self.bot.change_presence(activity=Activity(name=f'{self.bot.owner.name}\'s trades | V{__version__()}',
+                                                        type=ActivityType.watching))
+            print(f'{"-" * 30}\n{self.bot.user.name} is ready')
+            print(f'Send this id: "{self.bot.user.id}" to Gobot1234 to add your bot to the server to use the custom emojis')
+            print(f'This is: Version {__version__()}')
+            while self.bot.logged_on is False:
+                if self.bot.cli_login:
+                    await self.bot.owner.send('You aren\'t currently logged into your Steam account\nTo do that type '
+                                              'in your 2FA code into the console.')
+                await sleep(60)
+            print('-' * 30)
+            try:
+                self.bot.trades = int(open('trades.txt', 'r').read())
+                channel = self.bot.get_channel(int(open('channel.txt', 'r').read()))
+                if channel is not None:
+                    async for m in channel.history(limit=2):
+                        if m.author == self.bot.user:
+                            await m.delete()
+                    await channel.send('Finished restarting...', delete_after=10)
+                remove('channel.txt')
+                remove('trades.txt')
+            except FileNotFoundError:
+                pass
+            await self.bot.owner.send('I\'m online both Steam and Discord dealing with your Steam messages')
 
     @commands.Cog.listener()
     async def on_ready(self):
@@ -82,25 +75,11 @@ class Loader(commands.Cog):
 
     @commands.command()
     @commands.is_owner()
-    async def restart(self, ctx):
-        """Used to reset the bot
-
-        This won't work use this if you use `cli_login`, try not to use this as it can lead to memory shortages
-        due to not letting old processes die"""
-        if self.bot.cli_login:
-            await ctx.send('You really really shouldn\'t do this')
-        else:
-            open('channel.txt', 'w+').write(str(ctx.channel.id))
-            open('trades.txt', 'w+').write(str(self.bot.trades))
-            await ctx.send(f'**Restarting the bot** {ctx.author.mention}, don\'t use this often')
-            execv(executable, ['python'] + argv)
-
-    @commands.command()
-    @commands.is_owner()
     async def logout(self, ctx):
         """Log yourself out safely"""
         open('channel.txt', 'w+').write(str(ctx.channel.id))
         open('trades.txt', 'w+').write(str(self.bot.trades))
+        self.bot.session.close()
         await ctx.send('Logging out...')
         self.bot.client.logout()
         await self.bot.close()
